style(rental): rename rental usecases directory to useCases
